+++ avr/cores/tinymodern/WProgram.h
@@ -25,7 +25,7 @@ uint16_t makeWord(byte h, byte l);
 
 unsigned long pulseIn(uint8_t pin, uint8_t state, unsigned long timeout = 1000000L);
 
-void tone(uint8_t _pin, unsigned int frequency, unsigned long duration = 0);
+void tone(uint8_t _pin, unsigned long frequency, unsigned long duration = 0);
 void noTone(uint8_t _pin = 255);
 
 // WMath prototypes
